[#34] refactor: SubscribeCalendars와 MyCalendars가 중복되는 부분이 많아서 Calendars 컴포넌트 하나로 리팩토링
